Split walkability.rs into walkability/ subdirectory module
Break the 718-line walkability.rs into focused submodules:
- categories.rs (91 lines): WalkabilityCategory enum, classification helpers
- grid.rs (77 lines): WalkabilityGrid resource, Saveable impl
- scoring.rs (170 lines): distance decay, category scoring, update system
- tests.rs (363 lines): all unit tests
- mod.rs (60 lines): re-exports, plugin definition

All public items are re-exported from mod.rs so external imports are unchanged.

Part of #1389
